requirements.txt
# ML
ultralytics==8.4.33
opencv-python-headless==4.13.0.92
numpy==2.0.2
scipy==1.16.3

# API
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0

# Data
sqlalchemy[asyncio]==2.0.35
alembic==1.13.2
asyncpg==0.29.0

# GCP
google-cloud-storage==2.18.0
google-cloud-logging==3.11.0

# Testing
pytest==8.3.0
pytest-asyncio==0.25.3
httpx==0.27.0
aiosqlite==0.20.0